Fix two Windows-only leaks the merge brought across
CI caught both immediately, on every Linux and macOS job while all four Windows
ones passed -- which is the whole reason this went through a PR rather than
straight onto master. Neither is subtle once a POSIX compiler sees it; neither is
visible from Windows at all.

libs/codec/cellAdec.c used getenv with no <stdlib.h>. MSVC pulls the declaration
in transitively and clang and gcc do not, and an implicit declaration returns
int, so the pointer would have been truncated even where it did compile.

libs/input/cellPad.c declared a timestamp as ULONGLONG, a Windows typedef. The
same file already spells the same thing portably twenty lines earlier
((unsigned long long)GetTickCount64()), so this now matches its own convention.

+++ libs/input/cellPad.c
@@ -610,7 +610,7 @@ skip_inject: ;
      * Legit input simulation on the same footing as LBP_AUTOPRESS/PAD_STICK. */
     { static int s_sc = -1;
       static struct { double t; unsigned mask; } ev[32]; static int n_ev = 0;
-      static ULONGLONG t0 = 0;
+      static unsigned long long t0 = 0;   /* not ULONGLONG: this builds on POSIX too */
       if (s_sc < 0) {
           s_sc = 0;
           const char* e = getenv("PAD_SCRIPT");
